reformatted structure of Dataset
