+++ hypex/dataset/__init__.py
@@ -2,6 +2,7 @@
 This module defines data structures and roles used across the library for managing and manipulating experimental data.
 """
 
+from hypex.dataset.abstract import DatasetBase
 from hypex.dataset.dataset import Dataset, ExperimentData
 from hypex.dataset.roles import (
     Arg1Role,
@@ -39,4 +40,5 @@ __all__ = [
     "TempTargetRole",
     "ABCRole",
     "default_roles",
+    "DatasetBase",
 ]
